Invariant naming of kwbody methods
Requires Julia 1.4 to work.
This also reorganizes the file structure to make this package a little 
more maintainable.
