+++ src/SnoopCompile.jl
@@ -2,486 +2,18 @@ module SnoopCompile
 
 using Serialization
 
-export @snoopc
-if VERSION >= v"1.2.0-DEV.573"
-    export @snoopi
-
-    const __inf_timing__ = Tuple{Float64,Core.MethodInstance}[]
-
-    function typeinf_ext_timed(linfo::Core.MethodInstance, params::Core.Compiler.Params)
-        tstart = time()
-        ret = Core.Compiler.typeinf_ext(linfo, params)
-        tstop = time()
-        push!(__inf_timing__, (tstop-tstart, linfo))
-        return ret
-    end
-    function typeinf_ext_timed(linfo::Core.MethodInstance, world::UInt)
-        tstart = time()
-        ret = Core.Compiler.typeinf_ext(linfo, world)
-        tstop = time()
-        push!(__inf_timing__, (tstop-tstart, linfo))
-        return ret
-    end
-
-    @noinline start_timing() = ccall(:jl_set_typeinf_func, Cvoid, (Any,), typeinf_ext_timed)
-    @noinline stop_timing() = ccall(:jl_set_typeinf_func, Cvoid, (Any,), Core.Compiler.typeinf_ext)
-
-    function sort_timed_inf(tmin)
-        data = __inf_timing__
-        if tmin > 0
-            data = filter(tl->tl[1] >= tmin, data)
-        end
-        return sort(data; by=tl->tl[1])
-    end
-
-    macro snoopi(args...)
-        tmin = 0.0
-        if length(args) == 1
-            cmd = args[1]
-        elseif length(args) == 2
-            a = args[1]
-            if isa(a, Expr) && a.head == :(=) && a.args[1] == :tmin
-                tmin = a.args[2]
-                cmd = args[2]
-            else
-                error("unrecognized input ", a)
-            end
-        else
-            error("at most two arguments are supported")
-        end
-        quote
-            empty!(__inf_timing__)
-            start_timing()
-            try
-                $(esc(cmd))
-            finally
-                stop_timing()
-            end
-            $sort_timed_inf($tmin)
-        end
-    end
-
-    @doc """
-        inf_timing = @snoopi commands
-        inf_timing = @snoopi tmin=0.0 commands
-
-    Execute `commands` while snooping on inference. Returns an array of `(t, linfo)`
-    tuples, where `t` is the amount of time spent infering `linfo` (a `MethodInstance`).
-
-    Methods that take less time than `tmin` will not be reported.
-    """
-    :(SnoopCompile.@snoopi)
-
-end
-
-"""
-```
-@snoopc "compiledata.csv" begin
-    # Commands to execute, in a new process
-end
-```
-causes the julia compiler to log all functions compiled in the course
-of executing the commands to the file "compiledata.csv". This file
-can be used for the input to `SnoopCompile.read`.
-"""
-macro snoopc(flags, filename, commands)
-    return :(snoopc($(esc(flags)), $(esc(filename)), $(QuoteNode(commands))))
-end
-macro snoopc(filename, commands)
-    return :(snoopc(String[], $(esc(filename)), $(QuoteNode(commands))))
-end
-
-function snoopc(flags, filename, commands)
-    println("Launching new julia process to run commands...")
-    # addprocs will run the unmodified version of julia, so we
-    # launch it as a command.
-    code_object = """
-            using Serialization
-            while !eof(stdin)
-                Core.eval(Main, deserialize(stdin))
-            end
-            """
-    process = open(`$(Base.julia_cmd()) $flags --eval $code_object`, stdout, write=true)
-    serialize(process, quote
-        let io = open($filename, "w")
-            ccall(:jl_dump_compiles, Nothing, (Ptr{Nothing},), io.handle)
-            try
-                $commands
-            finally
-                ccall(:jl_dump_compiles, Nothing, (Ptr{Nothing},), C_NULL)
-                close(io)
-            end
-        end
-        exit()
-    end)
-    wait(process)
-    println("done.")
-    nothing
-end
-
-function split2(str, on)
-    i = findfirst(isequal(on), str)
-    i === nothing && return str, ""
-    return (SubString(str, firstindex(str), prevind(str, first(i))),
-            SubString(str, nextind(str, last(i))))
-end
-
-"""
-`SnoopCompile.read("compiledata.csv")` reads the log file produced by the compiler and returns the functions as a pair of arrays. The first array is the amount of time required to compile each function, the second is the corresponding function + types. The functions are sorted in order of increasing compilation time. (The time does not include the cost of nested compiles.)
-"""
-function read(filename)
-    times = Vector{UInt64}()
-    data = Vector{String}()
-    toplevel = false # workaround for Julia#22538
-    for line in eachline(filename)
-        if toplevel
-            if endswith(line, '"')
-                toplevel = false
-            end
-            continue
-        end
-        time, str = split2(line, '\t')
-        length(str) < 2 && continue
-        (str[1] == '"' && str[end] == '"') || continue
-        if startswith(str, """"<toplevel thunk> -> """)
-            # consume lines until we find the terminating " character
-            toplevel = true
-            continue
-        end
-        tm = tryparse(UInt64, time)
-        tm === nothing && continue
-        push!(times, tm)
-        push!(data, str[2:prevind(str, lastindex(str))])
-    end
-    # Save the most costly for last
-    p = sortperm(times)
-    return (permute!(times, p), permute!(data, p))
-end
-
-# pattern match on the known output of jl_static_show
-extract_topmod(e::QuoteNode) = extract_topmod(e.value)
-function extract_topmod(e)
-    Meta.isexpr(e, :.) &&
-        return extract_topmod(e.args[1])
-    Meta.isexpr(e, :call) && length(e.args) == 3 && e.args[1] == :getfield &&
-        return extract_topmod(e.args[2])
-    Meta.isexpr(e, :call) && length(e.args) == 2 && e.args[1] == :typeof &&
-        return extract_topmod(e.args[2])
-    #Meta.isexpr(e, :call) && length(e.args) == 2 && e.args[1] == :Symbol &&
-    #    return Symbol(e.args[2])
-    # parametrized anonymous functions
-    Meta.isexpr(e, :call) && e.args[1].args[1] == :getfield &&
-        return extract_topmod(e.args[1].args[2].args[2])
-    isa(e, Symbol) &&
-        return e
-    return :unknown
-end
-
 const anonrex = r"#{1,2}\d+#{1,2}\d+"         # detect anonymous functions
 const kwrex = r"^#kw##(.*)$|^#([^#]*)##kw$"   # detect keyword-supplying functions
+const kwbodyrex = r"^##(\w[^#]*)#\d+"         # detect keyword body methods
 const genrex = r"^##s\d+#\d+$"                # detect generators for @generated functions
 
-function parse_call(line; subst=Vector{Pair{String, String}}(), blacklist=String[])
-    match(anonrex, line) === nothing || return false, line, :unknown, ""
-    for (k, v) in subst
-        line = replace(line, k=>v)
-    end
-    if any(b -> occursin(b, line), blacklist)
-        println(line, " contains a blacklisted substring")
-        return false, line, :unknown, ""
-    end
-
-    curly = ex = Meta.parse(line, raise=false)
-    while Meta.isexpr(curly, :where)
-        curly = curly.args[1]
-    end
-    if !Meta.isexpr(curly, :curly)
-        @warn("failed parse of line: ", line)
-        return false, line, :unknown, ""
-    end
-    func = curly.args[2]
-    topmod = (func isa Expr ? extract_topmod(func) : :Main)
-
-    check = Meta.isexpr(func, :call) && length(func.args) == 3 && func.args[1] == :getfield
-    name = (check ? func.args[3].args[2] : "")
-
-    # make some substitutions to try to form a leaf types tuple
-    changed = false
-    for i in 3:length(curly.args)
-        e = curly.args[i]
-        if Meta.isexpr(e, :where) && length(e.args) == 2 && Meta.isexpr(e.args[1], :curly) && length(e.args[1].args) == 3 && e.args[1].args[1] === :Vararg
-            # Unwrap a Varargs argument, make it a single instance of the allowed Varargs type.
-            e = e.args[1].args[2]
-            curly.args[i] = e
-            changed = true
-        end
-        if e === :Function
-            # `Function` is not an allowable type for `precompile()`
-            e = :(typeof(identity))
-        elseif e === :Any
-            # `Any` is an abstract type, and can't be precompiled for. Replace it with something that can.
-            e = :Int
-        elseif e == :(Type{T} where T)
-            # Type is an abstract type; replace it with a concrete type.
-            e = :(Type{Int})
-        else
-            continue
-        end
-        curly.args[i] = e
-        changed = true
-    end
-    # In Julia 0.6, functions with symbolic names like `Base.:(+)` don't output correctly.
-    # So if we didn't change the arg list above, use the original string.
-    if changed
-        line = string(ex)
-    end
-    return true, line, topmod, name
-end
-
-function topmodule(mods)
-    function ischild(c, mod)
-        ok = false
-        pc = parentmodule(c)
-        while pc !== c
-            pc === Main && return false  # mostly important for passing the tests
-            if isdefined(mod, nameof(pc))
-                ok = true
-                break
-            end
-            c = pc
-        end
-        return ok
-    end
-
-    mods = collect(mods)
-    mod = first(mods)
-    for m in Iterators.drop(mods, 1)
-        # Easy cases
-        if isdefined(mod, nameof(m))
-        elseif isdefined(m, nameof(mod))
-            mod = m
-        else
-            # Check parents of each
-            if ischild(m, mod)
-            elseif ischild(mod, m)
-                mod = m
-            else
-                return nothing
-            end
-        end
-    end
-    return mod
-end
-
-function addmodules!(mods, parameters)
-    for p in parameters
-        if isa(p, DataType)
-            push!(mods, Base.moduleroot(p.name.module))
-            addmodules!(mods, p.parameters)
-        end
-    end
-    return mods
-end
-
-function methods_with_generators(m::Module)
-    meths = Method[]
-    for name in names(m; all=true)
-        isdefined(m, name) || continue
-        f = getfield(m, name)
-        if isa(f, Function)
-            for method in methods(f)
-                if isdefined(method, :generator)
-                    push!(meths, method)
-                end
-            end
-        end
-    end
-    return meths
-end
-
-function parcel(tinf::AbstractVector{Tuple{Float64,Core.MethodInstance}}; subst=Vector{Pair{String, String}}(), blacklist=String[])
-    pc = Dict{Symbol, Vector{String}}()      # output
-    modgens = Dict{Module, Vector{Method}}() # methods with generators in a module
-    mods = Set{Module}()                     # module of each parameter for a given method
-    for (t, mi) in reverse(tinf)
-        isdefined(mi, :specTypes) || continue
-        tt = mi.specTypes
-        m = mi.def
-        isa(m, Method) || continue
-        # Determine which module to assign this method to. All the types in the arguments
-        # need to be defined; we collect all the corresponding modules and assign it to the
-        # "topmost".
-        empty!(mods)
-        push!(mods, Base.moduleroot(m.module))
-        addmodules!(mods, tt.parameters)
-        topmod = topmodule(mods)
-        if topmod === nothing
-            @debug "Skipping $tt due to lack of a suitable top module"
-            continue
-        end
-        # If we haven't yet started the list for this module, initialize
-        topmodname = nameof(topmod)
-        if !haskey(pc, topmodname)
-            pc[topmodname] = String[]
-        end
-        # Create the string representation of the signature
-        # Use special care with keyword functions, anonymous functions
-        prefix = ""
-        p = tt.parameters[1]
-        mname, mmod = String(p.name.name), m.module   # m.name strips the kw identifier
-        mkw = match(kwrex, mname)
-        manon = match(anonrex, mname)
-        mgen = match(genrex, mname)
-        frepr = if mkw !== nothing
-            # Keyword function
-            fname = mkw.captures[1] === nothing ? mkw.captures[2] : mkw.captures[1]
-            "Core.kwftype(typeof($mmod.$fname))"
-        elseif manon !== nothing
-            # Anonymous function, wrap in an `isdefined`
-            prefix = "if isdefined($mmod, Symbol(\"$mname\")"
-            "getfield($mmod, Symbol(\"$mname\"))"  # this is universal, var is Julia 1.3+
-        elseif mgen !== nothing
-            # Generator for a @generated function
-            if !haskey(modgens, m.module)
-                callers = modgens[m.module] = methods_with_generators(m.module)
-            else
-                callers = modgens[m.module]
-            end
-            getgen = "nothing"
-            for caller in callers
-                if nameof(caller.generator.gen) == m.name
-                    getgen = "typeof(first(whichtt($(caller.sig))).generator.gen)"
-                    break
-                end
-            end
-            getgen
-        else
-            repr(p)
-        end
-        paramrepr = map(repr, Iterators.drop(tt.parameters, 1))
-        pushfirst!(paramrepr, frepr)
-        ttrepr = "Tuple{" * join(paramrepr, ',') * '}'
-        # Check that we parsed it correctly, and that all types are defined in the module
-        ttexpr = Meta.parse(ttrepr)
-        if mgen === nothing  # whichtt is not defined in topmod
-            try
-                Core.eval(topmod, ttexpr)
-            catch
-                @debug "Module $topmod: skipping $ttrepr due to eval failure"
-                continue
-            end
-        end
-        stmt = "precompile(" * ttrepr * ')'
-        if !isempty(prefix)
-            stmt = prefix * ' ' * stmt * " end"
-        end
-        push!(pc[topmodname], stmt)
-    end
-    return pc
-end
-
-"""
-`pc = parcel(calls; subst=[], blacklist=[])` assigns each compile statement to the module that owns the function. Perform string substitution via `subst=["Module1"=>"Module2"]`, and omit functions in particular modules with `blacklist=["Module3"]`. On output, `pc[:Module2]` contains all the precompiles assigned to `Module2`.
-
-Use `SnoopCompile.write(prefix, pc)` to generate a series of files in directory `prefix`, one file per module.
-"""
-function parcel(calls::AbstractVector{String}; subst=Vector{Pair{String, String}}(), blacklist=String[])
-    pc = Dict{Symbol, Vector{String}}()
-    for c in calls
-        local keep, pcstring, topmod
-        keep, pcstring, topmod, name = parse_call(c, subst=subst, blacklist=blacklist)
-        keep || continue
-        # Add to the appropriate dictionary
-        if !haskey(pc, topmod)
-            pc[topmod] = String[]
-        end
-        prefix = (isempty(name) ? "" : "isdefined($topmod, Symbol(\"$name\")) && ")
-        push!(pc[topmod], prefix * "precompile($pcstring)")
-    end
-    return pc
-end
-
-"""
-`pc = format_userimg(calls; subst=[], blacklist=[])` generates precompile directives intended for your base/userimg.jl script. Use `SnoopCompile.write(filename, pc)` to create a file that you can `include` into `userimg.jl`.
-"""
-function format_userimg(calls; subst=Vector{Pair{String, String}}(), blacklist=String[])
-    pc = Vector{String}()
-    for c in calls
-        keep, pcstring, topmod, name = parse_call(c, subst=subst, blacklist=blacklist)
-        keep || continue
-        prefix = (isempty(name) ? "" : "isdefined($topmod, Symbol(\"$name\")) && ")
-        push!(pc, prefix * "precompile($pcstring)")
-    end
-    return pc
-end
-
-# Write precompiles for userimg.jl
-function write(io::IO, pc::Vector)
-    for ln in pc
-        println(io, ln)
-    end
-end
-
-function write(filename::AbstractString, pc::Vector)
-    path, fn = splitdir(filename)
-    if !isdir(path)
-        mkpath(path)
-    end
-    open(filename, "w") do io
-        write(io, pc)
-    end
-    nothing
-end
-
-"""
-    write(prefix::AbstractString, pc::Dict; always::Bool = false)
-
-Write each modules' precompiles to a separate file.  If `always` is
-true, the generated function will always run the precompile statements
-when called, otherwise the statements will only be called during
-package precompilation.
-"""
-function write(prefix::AbstractString, pc::Dict; always::Bool = false)
-    if !isdir(prefix)
-        mkpath(prefix)
-    end
-    for (k, v) in pc
-        open(joinpath(prefix, "precompile_$k.jl"), "w") do io
-            println(io, """
-            # Like `which` but takes the full signature Tuple-type (including `typeof(f)`)
-            # Used to get the generator of @generated functions, if present
-            function whichtt(@nospecialize(tt))
-                m = ccall(:jl_gf_invoke_lookup, Any, (Any, UInt), tt, typemax(UInt))
-                m === nothing && return nothing
-                return m.func::Method
-            end
-            """)
-            println(io, "function _precompile_()")
-            !always && println(io, "    ccall(:jl_generating_output, Cint, ()) == 1 || return nothing")
-            for ln in v
-                println(io, "    ", ln)
-            end
-            println(io, "end")
-        end
-    end
-end
-
-macro snoop(args...)
-    @warn "@snoop is deprecated, use @snoopc instead"
-    :(@snoopc $(args...))
-end
-
 if VERSION >= v"1.2.0-DEV.573"
-    function __init__()
-        # typeinf_ext_timed must be compiled before it gets run
-        # We do this in __init__ to make sure it gets compiled to native code
-        # (the *.ji file stores only the inferred code)
-        precompile(typeinf_ext_timed, (Core.MethodInstance, Core.Compiler.Params))
-        precompile(typeinf_ext_timed, (Core.MethodInstance, UInt))
-        precompile(start_timing, ())
-        precompile(stop_timing, ())
-        nothing
-    end
+    include("snoopi.jl")
+    include("parcel_snoopi.jl")
 end
+include("snoopc.jl")
+include("parcel_snoopc.jl")
+
+include("write.jl")
 
 end # module
